Dino Game in Python
Setup Guide: https://itsourcecode.com/free-projects/python-projects/dino-game-in-python-with-source-code/

Dino Game is developed in Python Programming Language and it is a desktop application. This Dino Game in Python is free to download the open-source code and it is created for beginners who want to learn python. This project system used a Pygame and Random module. Pygame is a cross-platform set of Python modules designed for writing video games.